% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox_interval}
\alias{boxcox_interval}
\title{Extract the 1 - alpha confidence interval from a box-cox transformation.}
\usage{
boxcox_interval(boxcox_fit, alpha = 0.05)
}
\arguments{
\item{boxcox_fit}{The result of calling MASS::boxcox.}

\item{alpha}{The desired significance level.}
}
\value{
A list containing the maximum likelihood estimator of lambda
  (lambda_hat) and the (1 - alpha) CI for lambda.
}
\description{
Extract the 1 - $\alpha$ confidence interval from a box-cox transformation.
}
