% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{get_n_jobs}
\alias{get_n_jobs}
\title{Get number of jobs to run in parallel.}
\usage{
get_n_jobs(n_jobs)
}
\description{
Determines the number of jobs which are going to run in parallel.
This function is inspired by the python function in joblib.
}
