% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_pub}
\alias{scale_color_pub}
\title{Defines the color palette for the 'color' aesthetic used by the publication
theme.}
\usage{
scale_color_pub(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::discrete_scale}
}
\description{
Defines the color palette for the 'color' aesthetic used by the publication
theme.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = disp, color = factor(carb))) +
  geom_point() +
  scale_color_pub() +
  theme_pub()

}
