% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_pub}
\alias{scale_fill_pub}
\title{Defines the color palette for the 'fill' aesthetic used by the publication
theme.}
\usage{
scale_fill_pub(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::discrete_scale}
}
\description{
Defines the color palette for the 'fill' aesthetic used by the publication
theme.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = factor(carb), fill = factor(carb))) +
  geom_bar() +
  scale_fill_pub() +
  theme_pub()

}
