% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Publication Theme}
\usage{
theme_pub(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{The size of the fonts used throughout the theme.}

\item{base_family}{The font family used throughout the theme.}
}
\description{
This is a publication ready theme created by Koundinya Desiraju in the
following blog post: https://rpubs.com/Koundy/71792.
}
